requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
python-dateutil==2.9.0.post0
numpy==1.26.4
pandas==2.2.3
plotly==5.24.1
streamlit==1.39.0
requests==2.32.3
transformers==4.44.2
torch==2.4.1
sentence-transformers==3.1.1
faiss-cpu==1.8.0.post1
ollama==0.3.3
pytest==8.3.3
httpx==0.27.2